requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.6
chardet==4.0.0
discord.py==1.7.3
Faker==8.10.3
idna==3.2
mixer==7.1.2
multidict==5.1.0
pycparser==2.20
pygame==2.0.1
PyNaCl==1.4.0
pyscroll==2.19.2
python-dateutil==2.8.2
python-dotenv==0.18.0
PyTMX==3.25
six==1.16.0
text-unidecode==1.3
typing-extensions==3.10.0.0
yarl==1.6.3
wikipedia==1.4.0
beautifulsoup4==4.9.3
google==3.0.0